+++ 10_PROJET_WS/src/main/java/fr/adaming/service/IMarchandiseService.java
@@ -13,5 +13,5 @@ public Marchandise addMarchandise(Marchandise m);
 	
 	public List<Marchandise> getMarchandiseByCargaison (Cargaison c);
 	
-	public void deleteMarchandise (int id);
+	public void deleteMarchandise (long id);
 }
