+++ 10_PROJET_WS/src/main/java/fr/adaming/service/MarchandiseServiceImpl.java
@@ -37,7 +37,7 @@ public class MarchandiseServiceImpl implements IMarchandiseService{
 	}
 
 	@Override
-	public void deleteMarchandise(int id) {
+	public void deleteMarchandise(long id) {
 		marchDao.deleteMarchandise(id);
 		
 	}
